print dan delete memcache dari folder yg aktiv

+++ app/BPlugin/SempoaMaster/TestMemcache.php
@@ -44,10 +44,36 @@ class TestMemcache extends WebService
         }
 
     }
-    function getMemcacheKeys()
+
+    function printallKeys()
+    {
+        $memcache = new Memcache;
+        $memcache->connect('127.0.0.1', 11211)
+        or die ("Could not connect to memcache server");
+
+        $list = array();
+        $allSlabs = $memcache->getExtendedStats('slabs');
+        $items = $memcache->getExtendedStats('items');
+        foreach ($allSlabs as $server => $slabs) {
+            foreach ($slabs AS $slabId => $slabMeta) {
+                $cdump = $memcache->getExtendedStats('cachedump', (int)$slabId);
+                foreach ($cdump AS $keys => $arrVal) {
+                    if (!is_array($arrVal)) continue;
+                    foreach ($arrVal AS $k => $v) {
+
+                        echo $k . '<br>';
+                        pr($memcache->get($k));
+
+
+                    }
+                }
+            }
+        }
+    }
+
+    function getMemcacheKeysByMyFolder()
     {
 
-        global $domain;
         global $folder;
         $memcache = new Memcache;
         $memcache->connect('127.0.0.1', 11211)
@@ -62,68 +88,26 @@ class TestMemcache extends WebService
                 foreach ($cdump AS $keys => $arrVal) {
                     if (!is_array($arrVal)) continue;
                     foreach ($arrVal AS $k => $v) {
-                        echo $k .'<br>';
-                        pr($memcache->get($k));
+                        $check = strpos($k, $folder);
+                        if ($check) {
+                            echo $k . '<br>' . $check;
+                            pr($memcache->get($k));
+                        }
+
 
                     }
                 }
             }
         }
 
-
-
-//        $reg = new RegisterGuru();
-//        $reg->isInvoiceCreated(50);
-//        pr($reg);
-        echo "end";
-
-//
-//
-//
-//
-//        $biaya = Generic::getBiayaByJenis(KEY::$BIAYA_PENDAFTARAN_GURU, AccessRight::getMyOrgID());
-//        $guru = new SempoaGuruModel();
-//        $guru->getByID(45);
-//        pr($biaya);
-
-
-
-
-        die();
-        pr("Memcache");
-        pr($memcache->get("localhost:8888/sempoa_memcached//MatrixNilaiModel_c__2"));
-        $id =
-        $murid = $memcache->get("localhost:8888/sempoa_memcached//MuridModel_c__831");
-        pr($murid);
-//        pr($murid['nama_siswa']);
-        pr("Model");
-        $muridModel = new MuridModel();
-        $muridModel->getByID(831);
-
-        pr($muridModel);
-//        pr(get_called_class);
     }
 
-    function printMemcacheKey(){
 
+// Bahaya!!
+// Menghapus memcache di folder yg skrg lagi aktiv
+    function delMemcacheKeysByMyFolder()
+    {
 
-        $m = new Memcached();
-        $m->addServer('localhost', 11211);
-        $keys = $m->getAllKeys();
-        foreach ($keys as $index => $key) {
-            pr($key);
-//            if (strpos($key,$prefix) !== 0) {
-//                unset($keys[$index]);
-//            } else {
-////                $m->delete($key);
-//            }
-        }
-        return $keys;
-
-    }
-
-    function deletekeysbyindex() {
-        global $domain;
         global $folder;
         $memcache = new Memcache;
         $memcache->connect('127.0.0.1', 11211)
@@ -138,14 +122,19 @@ class TestMemcache extends WebService
                 foreach ($cdump AS $keys => $arrVal) {
                     if (!is_array($arrVal)) continue;
                     foreach ($arrVal AS $k => $v) {
-                        echo $k .'<br>';
+                        $check = strpos($k, $folder);
+                        if ($check) {
+                            echo $k . '<br>' . $check;
+                            $memcache->delete($k);
+                            pr($memcache->get($k));
+                        }
 
-                        $memcache->delete($k);
-                        pr($memcache->get($k));
 
                     }
                 }
             }
         }
+
     }
+
 }
